CI/CD for Lambda functions (#4)
* Add workflows

* CAT-318 CI/CD for Lambda functions (#3)

* Add gitignore template

* Add Makefile with terraform instructions

* Rename workflows files

* Relocate lambda functions

* Add terraform envs

* Add script glue job

* Add patterns for python

* Update for TF files

* Add compressed layers
