+++ .github/workflows/posts-batch-proccessing.yml
@@ -1,0 +1,67 @@
+name: Deploy posts-batch-processing
+
+on:
+  push:
+    branches:
+      - develop
+      - staging
+    paths:
+      - '../../lambdas/posts-batch-processing/**'
+  workflow_dispatch:
+    inputs:
+      environment:
+        description: 'Deployment environment'
+        required: false
+        default: ''
+      aws_region:
+        description: 'AWS Region for deployment'
+        required: false
+        default: 'us-west-1'
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+
+    env:
+      AWS_REGION: ${{ github.event.inputs.aws_region || 'us-west-1' }}
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v2.5.0
+
+    - name: Set up Python
+      uses: actions/setup-python@v2.3.0
+      with:
+        python-version: '3.11'
+
+    - name: Determine environment and region
+      id: setup-env
+      run: |
+        # Determine the environment
+        if [[ -n "${{ github.event.inputs.environment }}" ]]; then
+          echo "ENVIRONMENT=${{ github.event.inputs.environment }}" >> $GITHUB_ENV
+        elif [[ "${GITHUB_REF_NAME}" == "staging" ]]; then
+          echo "ENVIRONMENT=staging" >> $GITHUB_ENV
+        else
+          echo "ENVIRONMENT=develop" >> $GITHUB_ENV
+        fi
+  
+        # Determine the AWS region
+        if [[ -n "${{ github.event.inputs.aws_region }}" ]]; then
+          echo "AWS_REGION=${{ github.event.inputs.aws_region }}" >> $GITHUB_ENV
+        else
+          echo "AWS_REGION=us-west-1" >> $GITHUB_ENV
+          fi
+
+    - name: Package and deploy
+      env:
+        AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+        AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+        AWS_REGION: ${{ env.AWS_REGION }}
+      run: |
+        echo "Deploying to environment: $ENVIRONMENT in region: $AWS_REGION"
+        mkdir -p ../../compressed/
+        rm -f ../../compressed/${ENVIRONMENT}-postsbatchprocessing.zip
+        cd lambdas/posts-batch-processing
+        zip -r9 ../../compressed/${ENVIRONMENT}-postsbatchprocessing.zip .
+        aws lambda update-function-code --region ${AWS_REGION} --function-name ${ENVIRONMENT}-posts-batch-processing --zip-file fileb://../../compressed/${ENVIRONMENT}-postsbatchprocessing.zip .
