--- CreatorCatalyst-PostsProcessing/metrics_keys.json
@@ -1,4 +1,0 @@
-{
-    "youtube":["comments","views","likes"],
-    "instagram":["comments","likes","video_views","video_plays"]
-}
