--- CreatorCatalyst-ReportMetrics/platform.json
@@ -1,1 +1,0 @@
-{"tiktok":1,"instagram":2,"youtube":3,"facebook":4,"x":5}
