+++ envs/staging.tf
@@ -1,0 +1,1 @@
+environment = "staging"
