+++ lambdas/report-metrics/platform.json
@@ -1,0 +1,1 @@
+{"tiktok":1,"instagram":2,"youtube":3,"facebook":4,"x":5}
